refactor(organizations): rename postgres adapter to _postgresorganizationrepository
Align the organization Postgres adapter with the `_postgres<entity>repository.py`
naming convention already followed by the other adapters (and by its own test
file, `test_postgresorganizationrepository.py`).
