+++ api/infrastructure/postgres/__init__.py
@@ -2,7 +2,7 @@ from ._autocommitsession import AutocommitSession, TransactionRequiredError
 from ._postgresauthenticateduserquery import PostgresAuthenticatedUserQuery
 from ._postgreskeyrepository import PostgresKeyRepository
 from ._postgreslimitrepository import PostgresLimitRepository
-from ._postgresorganization import PostgresOrganizationRepository
+from ._postgresorganizationrepository import PostgresOrganizationRepository
 from ._postgrespermissionrepository import PostgresPermissionRepository
 from ._postgresproviderrepository import PostgresProviderRepository
 from ._postgresrolesrepository import PostgresRolesRepository
